Use custom JSONDecoder for Mlem backend
