+++ Mlem/App/Utility/Extensions/Content Models/Instance3+Extensions.swift
@@ -25,7 +25,7 @@ extension InstanceSummarySoftware {
     init(from software: SiteSoftware) {
         let type: InstanceSummarySoftwareType = switch software.type {
         case .lemmy: .lemmy
-        case .pieFed: .piefed
+        case .pieFed: .pieFed
         }
         
         self.init(
